refactor: update attrs
